LayoutTests/imported/w3c/web-platform-tests/2dcontext/imagebitmap/createImageBitmap-drawImage-expected.txt

CONSOLE MESSAGE: line 38: normal video: loadedmetadata
CONSOLE MESSAGE: line 38: normal video: loadeddata
CONSOLE MESSAGE: line 38: normal video: canplay
CONSOLE MESSAGE: line 41: loaded normal video
CONSOLE MESSAGE: line 54: got response
CONSOLE MESSAGE: line 59: creating video
CONSOLE MESSAGE: line 66: data video: loadedmetadata
CONSOLE MESSAGE: line 66: data video: loadeddata
CONSOLE MESSAGE: line 66: data video: canplay
CONSOLE MESSAGE: line 69: loaded data video

PASS createImageBitmap from an HTMLVideoElement, and drawImage on the created ImageBitmap 
PASS createImageBitmap from an HTMLVideoElement from a data URL, and drawImage on the created ImageBitmap 
